cross-platform: Compile into module blocks
 - Combines static libraries from sub projects. (Linux, macOS)
   ChakraCore consumer now links to 3 libs instead of 18.

 - Defines AutoSystemInfo etc. initializers builtin for ease of embedding
