+++ bin/GCStress/CMakeLists.txt
@@ -3,7 +3,7 @@ add_executable (GCStress
   RecyclerTestObject.cpp
   stdafx.cpp
   StubExternalApi.cpp
-    )
+  )
 
 include_directories(..)
 
@@ -13,46 +13,19 @@ target_include_directories (GCStress
   $<BUILD_INTERFACE:${ROOT_SOURCE_DIR}/lib/Common/Memory>  
   )
 
-set(LINK_FLAGS ${LINK_FLAGS} "-Wl,-whole-archive")
-
 if(CMAKE_SYSTEM_NAME STREQUAL Linux)
-  set(lib_target
-    -Wl,--no-undefined
-    -Wl,--start-group
-    )
-elseif(CMAKE_SYSTEM_NAME STREQUAL Darwin)
-  set(lib_target
-    -Wl,-undefined,error
-    )
-else()
-  message("This platform is not yet supported")
-endif() # Linux ?
-
-# make sure to include Common.Core before others
-# this will help linker on some platforms for correct
-# initialization order
-set(lib_target "${lib_target}"
-  Chakra.Common.Core
-  )
+  set(LINKER_START_GROUP -Wl,--start-group)
+  set(LINKER_END_GROUP -Wl,--end-group)
+endif()
 
 # common link deps
 set(lib_target "${lib_target}"
-  Chakra.Common.Memory
-  Chakra.Runtime.PlatformAgnostic
-  Chakra.Common.Common
-  Chakra.Common.DataStructures
-  Chakra.Common.Exceptions
+  -Wl,-undefined,error
+  ${LINKER_START_GROUP}
+  Chakra.Common.Core
+  Chakra.Pal
+  Chakra.Jsrt
+  ${LINKER_END_GROUP}
   )
-
-if(CMAKE_SYSTEM_NAME STREQUAL Linux)
-  set(lib_target "${lib_target}"
-    -Wl,--end-group
-    Chakra.Pal
-    )
-elseif(CMAKE_SYSTEM_NAME STREQUAL Darwin)
-  set(lib_target "${lib_target}"
-    Chakra.Pal
-    )
-endif() # Linux ?
 
 target_link_libraries (GCStress ${lib_target})
